Add target async handle to the task data structure
This commit adds a new opaque pointer to the task data structure. After
the task is executed, if the pointer contains a non-null address, the
task will be re-enqueued for execution on any hidden helper thread. This
allows a libomptarget function to be re-called for as long as it needs
to complete its asynchronous operations.