[LLVM][objcopy] Update Arm XFAIL in update section test
a1b4e13cff2a792571927ee1fc6eebb05e40fae9 updated this to use
the target= syntax.

However the triple for our Arm bots is usually like:
armv8l-unknown-linux-gnueabihf

With "eabihf" on the end. I assume before we just checked for
"linux-gnu" being in the triple at all but now it is a proper
regex match.

Add .* on the end to account for the ABI tag on the end.